import mac installer files

+++ Peakaboo_Packaging/build.rb
@@ -3,27 +3,60 @@
 #function to extract the jar file
 def setup
 
-	def doSetup(jarfile, path, resources)
+	def doSetup(jarfile, path, resources, unzip=true)
 
-		`rm -rf #{path}`
-		`unzip "#{jarfile}" -d #{path}`
+		if unzip
+			`rm -rf #{path}`
+			`unzip "#{jarfile}" -d "#{path}"`
+		else
+			`cp "#{jarfile}" "#{path}"`
+		end
+		
 		resources.each{|res|
-			`cp ./resources/#{res} #{path}`
+			source, target = res
+			`cp -f "./resources/#{source}" "#{target}"`
 		}
 	end
 
 	jarfile = `ls *.jar`.split("\n")[0]
 
-	#Deb Package
-	debpath = "./deb/3.0/usr/share/Peakaboo/"
-	resources = ["icon.png", "logo.png", "peakaboo.sh"]
+	 if jarfile == nil
+		puts "No JAR file found."
+		exit(1)
+	 end
+	 
+	 `mv -f "./#{jarfile}" "./Peakaboo.jar"`
+	 
+	 jarfile = "Peakaboo.jar"
 
-	doSetup(jarfile, debpath, resources)
+	#Deb Package
+	dapppath = "./deb/3.0/usr/share/Peakaboo/"
+	dbinpath = "./deb/3.0/usr/bin/"
+	
+	resources = [["shared/icon.png", dapppath], ["shared/logo.png", dapppath], ["linux/peakaboo", dbinpath]]
+
+	doSetup(jarfile, dapppath, resources)
+
+
+
 
 	#Win32 Package
 	winpath = "./windows/Peakaboo/"
-	resources = ["Logo.ico", "peakaboo.vbs"]
+	
+	resources = [["windows/Logo.ico", winpath], ["windows/peakaboo.vbs", winpath]]
+	
 	doSetup(jarfile, winpath, resources)
+	
+	
+	
+	
+	#Mac OS Package
+	macpath = "./mac/Peakaboo.app/Contents/MacOS"
+	macrespath = "./mac/Peakaboo.app/Contents/Resources"
+	
+	resources = [["mac/peakaboo", macpath], ["mac/peakaboo.icns", macrespath]]
+	
+	doSetup(jarfile, macpath, resources, false)
 
 end
 
@@ -43,6 +76,10 @@ puts "Building Debian Package..."
 puts "DONE\n\n"
 
 puts "Building Windows Package..."
+`rm -rf ./windows/_win32/_win32/*.exe`
 `cd ./windows/_win32/ && ./buildWindowsInstaller.sh`
-`cp "./windows/_win32/_win32/Peakaboo 3 Setup.exe" ./build/`
+`cp "./windows/_win32/_win32/*.exe" ./build/`
 puts "DONE\n\n"
+
+puts "Building Mac Package..."
+`cp -rf ./mac/Peakaboo.app/ ./build/`
